Move `templating` package outside of `reconcilier` 🍵
It's completely independent of the reconcilier or the apis 👼
