+++ pkg/apis/pipeline/v1alpha1/pipeline_validation.go
@@ -20,7 +20,7 @@ import (
 	"fmt"
 
 	"github.com/knative/build-pipeline/pkg/list"
-	"github.com/knative/build-pipeline/pkg/reconciler/v1alpha1/templating"
+	"github.com/knative/build-pipeline/pkg/templating"
 	"github.com/knative/pkg/apis"
 	"k8s.io/apimachinery/pkg/api/equality"
 )
