+++ pkg/apis/pipeline/v1alpha1/task_validation.go
@@ -20,7 +20,7 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/knative/build-pipeline/pkg/reconciler/v1alpha1/templating"
+	"github.com/knative/build-pipeline/pkg/templating"
 	"github.com/knative/pkg/apis"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/equality"
