+++ pkg/reconciler/v1alpha1/pipelinerun/resources/apply.go
@@ -17,7 +17,7 @@ import (
 	"fmt"
 
 	"github.com/knative/build-pipeline/pkg/apis/pipeline/v1alpha1"
-	"github.com/knative/build-pipeline/pkg/reconciler/v1alpha1/templating"
+	"github.com/knative/build-pipeline/pkg/templating"
 )
 
 // ApplyParameters applies the params from a PipelineRun.Params to a PipelineSpec.
