+++ pkg/reconciler/v1alpha1/taskrun/resources/apply.go
@@ -20,7 +20,7 @@ import (
 	"fmt"
 
 	"github.com/knative/build-pipeline/pkg/apis/pipeline/v1alpha1"
-	"github.com/knative/build-pipeline/pkg/reconciler/v1alpha1/templating"
+	"github.com/knative/build-pipeline/pkg/templating"
 	buildv1alpha1 "github.com/knative/build/pkg/apis/build/v1alpha1"
 )
